Add "resume" option
It happens that due to connection issues or other conditions executeDqChecks fails and progress is lost. Resume option allows to recover progress of previous execution.

Once certain check is processed without error, it's saved to file corresponding to check description. If overall execution completes without errors these intermediate files are deleted.

When resume option is set to TRUE and intermediate results for certain check exist, processing of the check is skipped and saved results are used instead.