Fix missing pull() and wrong launchLogViewer() file.path() argument
Update the checksToExclude by changing the select() to a pull() to make sure we return the vector only rather than a column. Add explicit namespacing and call the library(dplyr) at the top of the commented code. Update the inspect logs code to correct the file.path() argument to include the cdmSourceName only in the .txt file name.